+++ render/template/template.go
@@ -1,19 +1,21 @@
 package template
 
 import (
+	"compress/gzip"
 	"fmt"
+	"io"
 
 	"github.com/kataras/iris/config"
 	"github.com/kataras/iris/context"
 	"github.com/kataras/iris/render/template/engine/html"
 	"github.com/kataras/iris/render/template/engine/pongo"
+	"github.com/kataras/iris/utils"
 )
 
 type (
 	Engine interface {
 		BuildTemplates() error
-		Execute(ctx context.IContext, name string, binding interface{}, layout string) error
-		ExecuteGzip(ctx context.IContext, name string, binding interface{}, layout string) error
+		ExecuteWriter(out io.Writer, name string, binding interface{}, layout string) error
 	}
 
 	Template struct {
@@ -23,6 +25,7 @@ type (
 		Gzip          bool
 		ContentType   string
 		Layout        string
+		buffer        *utils.BufferPool // this is used only for RenderString
 	}
 )
 
@@ -52,11 +55,12 @@ func New(c config.Template) *Template {
 		Gzip:          c.Gzip,
 		ContentType:   compiledContentType,
 		Layout:        c.Layout,
+		buffer:        utils.NewBufferPool(64),
 	}
 
 }
 
-func (t *Template) Render(ctx context.IContext, name string, bindings interface{}, layout ...string) error {
+func (t *Template) Render(ctx context.IContext, name string, binding interface{}, layout ...string) (err error) {
 
 	if t == nil { // No engine was given but .Render was called
 		ctx.WriteHTML(403, "<b> Iris </b> <br/> Templates are disabled via config.NoEngine, check your iris' configuration please.")
@@ -65,11 +69,11 @@ func (t *Template) Render(ctx context.IContext, name string, bindings interface{
 
 	// build templates again on each render if IsDevelopment.
 	if t.IsDevelopment {
-		if err := t.Engine.BuildTemplates(); err != nil {
-			return err
+		if err = t.Engine.BuildTemplates(); err != nil {
+			return
 		}
 	}
-	ctx.GetRequestCtx().Response.Header.Set("Content-Type", t.ContentType)
+
 	// I don't like this, something feels wrong
 	_layout := ""
 	if len(layout) > 0 {
@@ -80,11 +84,52 @@ func (t *Template) Render(ctx context.IContext, name string, bindings interface{
 	}
 
 	//
+	ctx.GetRequestCtx().Response.Header.Set("Content-Type", t.ContentType)
+
+	var out io.Writer
 
 	if t.Gzip {
-		return t.Engine.ExecuteGzip(ctx, name, bindings, _layout)
+		out = gzip.NewWriter(ctx.GetRequestCtx().Response.BodyWriter()) // maybe this should have its own pool but how to write after? no .WriteTo method here, I will see it someday
+		defer out.(*gzip.Writer).Close()                                // here we need to close the gzip writer, after we don't do that because it is the context's body writer, carefuly kataras!
+		ctx.GetRequestCtx().Response.Header.Add("Content-Encoding", "gzip")
+	} else {
+		out = ctx.GetRequestCtx().Response.BodyWriter()
 	}
 
-	return t.Engine.Execute(ctx, name, bindings, _layout)
+	err = t.Engine.ExecuteWriter(out, name, binding, _layout)
 
+	return
+}
+
+func (t *Template) RenderString(name string, binding interface{}, layout ...string) (result string, err error) {
+
+	if t == nil { // No engine was given but .Render was called
+		err = fmt.Errorf("[IRIS TEMPLATES] Templates are disabled via config.NoEngine, check your iris' configuration please.\n")
+		return
+	}
+
+	// build templates again on each render if IsDevelopment.
+	if t.IsDevelopment {
+		if err = t.Engine.BuildTemplates(); err != nil {
+			return
+		}
+	}
+
+	// I don't like this, something feels wrong
+	_layout := ""
+	if len(layout) > 0 {
+		_layout = layout[0]
+	}
+	if _layout == "" {
+		_layout = t.Layout
+	}
+
+	out := t.buffer.Get()
+	// if we have problems later consider that -> out.Reset()
+	defer t.buffer.Put(out)
+	err = t.Engine.ExecuteWriter(out, name, binding, _layout)
+	if err == nil {
+		result = out.String()
+	}
+	return
 }